+++ launchDaemonManagement.zsh
@@ -18,17 +18,6 @@
 # to comply with DDM-enforced macOS update deadlines.
 #
 ####################################################################################################
-#
-# HISTORY
-#
-# Version 2.3.0, 19-Jan-2026, Dan K. Snelson (@dan-snelson)
-# - Refactored Update Required logic to address Feature Request #55
-# - Updated "Organization Variables" (i.e., removed redundant variable declarations)
-# - Refactored `OrganizationOverlayIconURL` logic to address Bug Report #56 (thanks, @walkintom!)
-# - Added hard-coded `disableButton2InsteadOfHide` variable to disable `button2`, instead of only hiding it (Inspired by Bug Report #58, thanks @ScottEKendall!)
-# - Replaced `defaults read` with PlistBuddy for prefs (Pull Request #61; thanks, @huexley!)
-#
-####################################################################################################
 
 
 
@@ -41,7 +30,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.3.0"
+scriptVersion="2.4.0"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"